Refactor App component and add useTasks hook
